backend_fastapi/pyproject.toml
[tool.poetry]
name = "backend-fastapi"
version = "0.1.0"
description = ""
authors = ["Andy <mazhiyuannn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
aiomysql = "^0.2.0"
sqlalchemy = "^2.0.30"
databases = "^0.9.0"
python-dotenv = "^1.0.1"
cryptography = "^42.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
